Rename `utiles` module to `utilities`
